requirements.txt
onnxruntime
rembg
pillow
matplotlib
numpy